fix tests

+++ pkg/ledger/executor_test.go
@@ -133,7 +133,7 @@ func TestVariables(t *testing.T) {
 		require.NoError(t, err)
 
 		b := user.Balances["CAD/2"]
-		assert.Equalf(t, 42, b,
+		assert.Equalf(t, int64(42), b,
 			"wrong CAD/2 balance for account user:042, expected: %d got: %d",
 			42, b,
 		)
